backend/requirements.txt
flask>=3.0
psycopg2-binary>=2.9
python-dotenv>=1.0
flask-cors>=4.0
gunicorn>=21.0
